requirements.txt
Django==1.11.4
django-ckeditor==5.3.0
django-debug-toolbar==1.8
django-js-asset==0.1.1
django-mptt==0.8.7
django-suit==0.2.25
mysqlclient==1.3.10
olefile==0.44
Pillow==4.2.1
pytz==2017.2
sqlparse==0.2.3
